Added Volrath's Shapeshifter.
The Graveyard was modified to become its own class, as the top card of
your graveyard matters for this card, and graveyard order matters for
some other old card. Note that due how Volrath's Shapeshifter's ability
works, it is applied in layer 3, after control effects. This is so it
can look at it's current controllers graveyard, instead of it's owner,
as it would if the copy effect was at layer 1.